// src/types/modules.ts
/**
 * ModuleTypes enum
 * 
 * Represents the different modules that can be enabled for a DAO.
 * Each module extends the functionality of a DAO with specific features.
 * 
 * This enum should match the Modules enum defined in the API.
 */
export enum ModuleTypes {
  PODS = "PODS",
  PROOF_OF_LOVE = "PROOF_OF_LOVE"
} 